refactor(daemon): Split up update
No code changes here, just some re-organization:

The `Updateable[...]` class definitions are pulled out of
`daemon/update.py` and into a new submodule `daemon.update`.
The remainder of `update.py`, containing the `alpenhornd` main
loop and associated bits, is renamed to `daemon.main`.

This is forward planning for the changes that will be needed to
support containerising, if that ever gets implemented.
