+++ alpenhorn/daemon/__init__.py
@@ -2,7 +2,7 @@
 
 from .. import __version__
 from .entry import entry
-from .update import host
+from .main import host
 
 # These classes are used by extensions, so let's import them into the daemon
 # base
