+++ tests/daemon/update/test_group.py
